Expand Phase 4 with user API, backend abstraction, SV convenience, and fix AXI-Lite BFM
Major additions:
- specs 08 (backend abstraction), 09 (user API), 10 (SV convenience interfaces)
- vten/functional.py: Level 3 API (run_kernel/KernelExecutor)
- vten/runtime: ExecutionContext.run() orchestration, RuntimeEngine enhancements
- vten/cli: build/run/init reworked for kernel-centric project layout
- vten/codegen: wrapper.sv.j2, axilite_ctrl.sv.j2, project_setup.tcl.j2 templates
- vten_sv: AXI interface modules (axilite_if, aximm_if, axis_if)
- examples/passthrough: kernel-per-directory layout, stream_dma examples

AXI-Lite BFM fixes (vten_bfm_axilite.sv):
- Registered idle signal to avoid xsim continuous-assign issue
- Added aw_accepted/w_accepted handshake tracking with default deassertions
- Fixed POLL_REG m_rready timing: drive rready from first AR cycle

Test driver fix (axilite_driver.cpp):
- Reordered R handshake clearance before AR acceptance in slave_respond()
  to prevent stale R completion from clobbering new AR request's ar_pending

All SV BFMs updated: removed $display statements, registered idle signals,
proper handshake phase tracking.

1473/1473 tests passing.

+++ vten_sv/vten_shm_bridge.c
@@ -325,17 +325,17 @@ void vten_read_command_deps(int cmd_id,
     uint16_t* deps  = (uint16_t*)(slot + 0x2C);
     uint16_t* cdeps = (uint16_t*)(slot + 0x34);
 
-    /* svOpenArrayHandle → contiguous uint16_t array */
-    uint16_t* dep_out  = (uint16_t*)svGetArrayPtr((svOpenArrayHandle)dep_ids_handle);
-    uint16_t* cdep_out = (uint16_t*)svGetArrayPtr((svOpenArrayHandle)cdep_ids_handle);
+    /* DPI-C int open array: svGetArrayPtr returns int* for 'output int arr[]' */
+    int* dep_out  = (int*)svGetArrayPtr((svOpenArrayHandle)dep_ids_handle);
+    int* cdep_out = (int*)svGetArrayPtr((svOpenArrayHandle)cdep_ids_handle);
 
     if (dep_out != NULL) {
         for (int i = 0; i < 4; i++)
-            dep_out[i] = deps[i];
+            dep_out[i] = (int)deps[i];
     }
     if (cdep_out != NULL) {
         for (int i = 0; i < 4; i++)
-            cdep_out[i] = cdeps[i];
+            cdep_out[i] = (int)cdeps[i];
     }
 }
